Iniciando o componente Share e Fav Buttons
