+++ src/pages/DetalhesBebida.js
@@ -1,12 +1,7 @@
 import React, { useEffect, useContext } from 'react';
-import { useLocation } from 'react-router-dom';
 import DetailsContext from '../context/DetailsContext';
+import ShareFavButtons from '../components/ShareFavButtons';
 import './Detalhes.css';
-
-const shareUrl = (location) => {
-  window.navigator.clipboard.writeText(`http://localhost:3000${location}`);
-  alert('Link copiado!');
-};
 
 const getIngredients = (obj) => {
   let ingredientsArray = []; let measureArray = [];
@@ -59,11 +54,10 @@ const recomendations = (obj) => {
 };
 
 const Detalhes = () => {
-  const location = useLocation();
+  const { fetchMeals, meals, mealsOk, location } = useContext(DetailsContext);
+  const { fetchDrink, drink, drinkOk, copyUrl } = useContext(DetailsContext);
   const address = location.pathname;
   const id = address.slice(9, address.length);
-  const { fetchMeals, meals, mealsOk } = useContext(DetailsContext);
-  const { fetchDrink, drink, drinkOk } = useContext(DetailsContext);
 
   useEffect(() => {
     fetchMeals();
@@ -77,13 +71,8 @@ const Detalhes = () => {
         <img data-testid="recipe-photo" src={drink.strDrinkThumb} width="360px" alt="Recipe" />
         <p data-testid="recipe-title">{drink.strDrink}</p>
         <p data-testid="recipe-category">{drink.strAlcoholic}</p>
-        <div>
-          <button
-            data-testid="share-btn" onClick={() => shareUrl(location.pathname)}
-          >Share
-          </button>
-          <button data-testid="favorite-btn">Favorite</button>
-        </div>;
+        <ShareFavButtons />
+        {copyUrl && <span>Link copiado!</span>}
         <p>{drink.strCategory}</p>
         <div><span>Ingredients</span>
           {
